# web/tooling/openapi/workforce.openapi.json
{
  "openapi": "3.1.1",
  "info": {
    "title": "Workforce.Host.Api | v1",
    "version": "1.0.0"
  },
  "servers": [
    {
      "url": "http://127.0.0.1:5041/"
    }
  ],
  "paths": {
...
    "/api/v1/recruitment/interviews": {
      "get": {
        "tags": [
          "RecruitmentInterviews"
        ],
        "parameters": [
          {
            "name": "startUtc",
            "in": "query",
            "schema": {
              "type": "string",
              "format": "date-time"
            }
          },
          {
            "name": "endUtc",
            "in": "query",
            "schema": {
              "type": "string",
              "format": "date-time"
            }
          },
          {
            "name": "interviewerUserId",
            "in": "query",
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Interview"
                  }
                }
              },
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Interview"
                  }
                }
              },
              "text/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Interview"
                  }
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "RecruitmentInterviews"
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ScheduleInterviewRequest"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/ScheduleInterviewRequest"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/ScheduleInterviewRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Created",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/Interview"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Interview"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/Interview"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/recruitment/interviews/{id}": {
      "get": {
        "tags": [
          "RecruitmentInterviews"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/Interview"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Interview"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/Interview"
                }
              }
            }
          },
          "404": {
            "description": "Not Found",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/recruitment/interviews/{id}/reschedule": {
      "post": {
        "tags": [
          "RecruitmentInterviews"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RescheduleInterviewRequest"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/RescheduleInterviewRequest"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/RescheduleInterviewRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/Interview"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Interview"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/Interview"
                }
              }
            }
          },
          "409": {
            "description": "Conflict",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/recruitment/interviews/{id}/complete": {
      "post": {
        "tags": [
          "RecruitmentInterviews"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ConcurrencyActionRequest"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/ConcurrencyActionRequest"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/ConcurrencyActionRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/Interview"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Interview"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/Interview"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/recruitment/interviews/{id}/cancel": {
      "post": {
        "tags": [
          "RecruitmentInterviews"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ConcurrencyActionRequest"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/ConcurrencyActionRequest"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/ConcurrencyActionRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/Interview"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Interview"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/Interview"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/recruitment/interviews/{id}/scorecard": {
      "post": {
        "tags": [
          "RecruitmentInterviews"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SubmitScorecardRequest"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/SubmitScorecardRequest"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/SubmitScorecardRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/ScorecardSubmission"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ScorecardSubmission"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/ScorecardSubmission"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/recruitment/interviews/{id}/scorecards": {
      "get": {
        "tags": [
          "RecruitmentInterviews"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/ScorecardSubmission"
                  }
                }
              },
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/ScorecardSubmission"
                  }
                }
              },
              "text/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/ScorecardSubmission"
                  }
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "ConcurrencyActionRequest": {
        "required": [
          "rowVersion"
        ],
        "type": "object",
        "properties": {
          "rowVersion": {
            "pattern": "^-?(?:0|[1-9]\\d*)$",
            "type": [
              "integer",
              "string"
            ],
            "format": "uint32"
          }
        }
      },
...
      "Interview": {
        "required": [
          "id",
          "tenantId",
          "applicationId",
          "stageId",
          "title",
          "interviewType",
          "scheduledStartUtc",
          "scheduledEndUtc",
          "timezone"
        ],
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "tenantId": {
            "$ref": "#/components/schemas/TenantId"
          },
          "applicationId": {
            "type": "string",
            "format": "uuid"
          },
          "stageId": {
            "type": "string",
            "format": "uuid"
          },
          "title": {
            "type": "string"
          },
          "interviewType": {
            "$ref": "#/components/schemas/InterviewType"
          },
          "scheduledStartUtc": {
            "type": "string",
            "format": "date-time"
          },
          "scheduledEndUtc": {
            "type": "string",
            "format": "date-time"
          },
          "timezone": {
            "type": "string"
          },
          "locationOrMeetingUrl": {
            "type": [
              "null",
              "string"
            ]
          },
          "status": {
            "$ref": "#/components/schemas/InterviewStatus"
          },
          "interviewKitJson": {
            "type": [
              "null",
              "string"
            ]
          },
          "createdAtUtc": {
            "type": "string",
            "format": "date-time"
          },
          "rowVersion": {
            "pattern": "^-?(?:0|[1-9]\\d*)$",
            "type": [
              "integer",
              "string"
            ],
            "format": "uint32"
          },
          "participants": {
            "type": [
              "null",
              "array"
            ],
            "items": {
              "$ref": "#/components/schemas/InterviewParticipant"
            }
          },
          "scorecards": {
            "type": [
              "null",
              "array"
            ],
            "items": {
              "$ref": "#/components/schemas/ScorecardSubmission"
            }
          }
        }
      },
      "InterviewerRole": {
        "type": "integer"
      },
      "InterviewParticipant": {
        "required": [
          "id",
          "interviewId",
          "interviewerUserId",
          "role"
        ],
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "interviewId": {
            "type": "string",
            "format": "uuid"
          },
          "interviewerUserId": {
            "type": "string",
            "format": "uuid"
          },
          "role": {
            "$ref": "#/components/schemas/InterviewerRole"
          },
          "isRequired": {
            "type": "boolean",
            "default": true
          }
        }
      },
      "InterviewParticipantDto": {
        "required": [
          "interviewerUserId",
          "role",
          "isRequired"
        ],
        "type": "object",
        "properties": {
          "interviewerUserId": {
            "type": "string",
            "format": "uuid"
          },
          "role": {
            "$ref": "#/components/schemas/InterviewerRole"
          },
          "isRequired": {
            "type": "boolean"
          }
        }
      },
      "InterviewStatus": {
        "type": "integer"
      },
      "InterviewType": {
        "type": "integer"
      },
...
      "ProblemDetails": {
        "type": "object",
        "properties": {
          "type": {
            "type": [
              "null",
              "string"
            ]
          },
          "title": {
            "type": [
              "null",
              "string"
            ]
          },
          "status": {
            "pattern": "^-?(?:0|[1-9]\\d*)$",
            "type": [
              "null",
              "integer",
              "string"
            ],
            "format": "int32"
          },
          "detail": {
            "type": [
              "null",
              "string"
            ]
          },
          "instance": {
            "type": [
              "null",
              "string"
            ]
          }
        }
      },
...
      "RescheduleInterviewRequest": {
        "required": [
          "newStartUtc",
          "newEndUtc",
          "timezone",
          "locationOrMeetingUrl",
          "expectedRowVersion"
        ],
        "type": "object",
        "properties": {
          "newStartUtc": {
            "type": "string",
            "format": "date-time"
          },
          "newEndUtc": {
            "type": "string",
            "format": "date-time"
          },
          "timezone": {
            "type": "string"
          },
          "locationOrMeetingUrl": {
            "type": [
              "null",
              "string"
            ]
          },
          "expectedRowVersion": {
            "pattern": "^-?(?:0|[1-9]\\d*)$",
            "type": [
              "integer",
              "string"
            ],
            "format": "uint32"
          }
        }
      },
...
      "ScheduleInterviewRequest": {
        "required": [
          "applicationId",
          "stageId",
          "title",
          "interviewType",
          "scheduledStartUtc",
          "scheduledEndUtc",
          "timezone",
          "locationOrMeetingUrl",
          "interviewKitJson",
          "participants"
        ],
        "type": "object",
        "properties": {
          "applicationId": {
            "type": "string",
            "format": "uuid"
          },
          "stageId": {
            "type": "string",
            "format": "uuid"
          },
          "title": {
            "type": "string"
          },
          "interviewType": {
            "$ref": "#/components/schemas/InterviewType"
          },
          "scheduledStartUtc": {
            "type": "string",
            "format": "date-time"
          },
          "scheduledEndUtc": {
            "type": "string",
            "format": "date-time"
          },
          "timezone": {
            "type": "string"
          },
          "locationOrMeetingUrl": {
            "type": [
              "null",
              "string"
            ]
          },
          "interviewKitJson": {
            "type": [
              "null",
              "string"
            ]
          },
          "participants": {
            "type": [
              "null",
              "array"
            ],
            "items": {
              "$ref": "#/components/schemas/InterviewParticipantDto"
            }
          }
        }
      },
      "ScorecardRecommendation": {
        "type": "integer"
      },
      "ScorecardSubmission": {
        "required": [
          "id",
          "interviewId",
          "applicationId",
          "interviewerUserId",
          "ratingsJson",
          "strengths",
          "concerns",
          "recommendation"
        ],
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "interviewId": {
            "type": "string",
            "format": "uuid"
          },
          "applicationId": {
            "type": "string",
            "format": "uuid"
          },
          "interviewerUserId": {
            "type": "string",
            "format": "uuid"
          },
          "ratingsJson": {
            "type": "string"
          },
          "strengths": {
            "type": [
              "null",
              "string"
            ]
          },
          "concerns": {
            "type": [
              "null",
              "string"
            ]
          },
          "recommendation": {
            "$ref": "#/components/schemas/ScorecardRecommendation"
          },
          "isFinalized": {
            "type": "boolean",
            "default": true
          },
          "submittedAtUtc": {
            "type": "string",
            "format": "date-time"
          },
          "rowVersion": {
            "pattern": "^-?(?:0|[1-9]\\d*)$",
            "type": [
              "integer",
              "string"
            ],
            "format": "uint32"
          }
        }
      },
...
      "SubmitScorecardRequest": {
        "required": [
          "ratingsJson",
          "strengths",
          "concerns",
          "recommendation",
          "expectedRowVersion"
        ],
        "type": "object",
        "properties": {
          "ratingsJson": {
            "type": "string"
          },
          "strengths": {
            "type": [
              "null",
              "string"
            ]
          },
          "concerns": {
            "type": [
              "null",
              "string"
            ]
          },
          "recommendation": {
            "$ref": "#/components/schemas/ScorecardRecommendation"
          },
          "expectedRowVersion": {
            "pattern": "^-?(?:0|[1-9]\\d*)$",
            "type": [
              "integer",
              "string"
            ],
            "format": "uint32"
          }
        }
      },
      "TenantId": {
        "type": "object",
        "properties": {
          "value": {
            "type": "string",
            "format": "uuid"
          }
        }
      },
...
    }
  },
...
}